Add llvm/projects/LLVMBuild.txt
Without this file I would get a build error when the file
LibraryDependencies.inc would not be found. This file is not present in
this git repo, but it is in the 3.3 source tarball from llvm.org. I
think the reason it wasn't included is that llvm/.gitignore ignores it.
However it specifically does include the other files that exist by
default in that directory.